chore: move utils in resolver.go
